% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_hierarchy.R
\name{cell_type_fate_similartiy}
\alias{cell_type_fate_similartiy}
\title{Calculate cell fate similarity among all cell types}
\usage{
cell_type_fate_similartiy(
  data,
  idx = "celltype",
  method = "spearman",
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{data.frame, indicating lineage tracing data, first column should
be lineage tracing barcodes, second column should be related cell type}

\item{idx}{column name of cell type}

\item{method}{character, indicating which correlation coefficient
(or covariance) is to be computed. One of "pearson", "kendall",
or "spearman (default)": can be abbreviated.}

\item{plot}{whether plot the heatmap}

\item{...}{other pheatmap parameters}
}
\value{
barcodes correlation between cell type
}
\description{
Calculate cell fate similarity among all cell types
}
\examples{
cell_type_fate_similartiy(fate_test)
}
