% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppatt.R
\name{CFMexplorer}
\alias{CFMexplorer}
\title{Shiny app for interactive visualization of expression patterns related to cell fate map}
\usage{
CFMexplorer(basis, init = 3, template)
}
\arguments{
\item{basis}{The basis component of a factorization of the spatial expression matrix}

\item{init}{initial value for thresholding slider}

\item{template}{matrix of 0 and 1, 1 indicating which parts of image rectangle are represented in the basis}
}
\description{
Shiny app for interactive visualization of expression patterns related to cell fate map
}
\note{
A spatial expression matrix has rows corresponding to a mapping of a grid, 
and columns corresponding to a gene; see \code{data(expressionPatterns)} 
for an example.  A reexpression of such a matrix using non-negative matrix
factorization is described in Wu et al., 
PNAS, 2016 \url{http://www.pnas.org/content/113/16/4290.full}.  The basis
matrix of this factorization yields what Wu et al. call 'principal patterns'.
To visualize these patterns in the context of the cell fate map, we take
the convex hulls of points having pattern weights exceeding a given
threshold.
}
