% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{DmLandmarks}
\alias{DmLandmarks}
\title{produce a data.frame with approximate landmarks of structures from Campos-Ortega and Hartenstein 1986 for the 405 point elliptical template}
\usage{
DmLandmarks()
}
\description{
produce a data.frame with approximate landmarks of structures from Campos-Ortega and Hartenstein 1986 for the 405 point elliptical template
}
