% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmMapTerms}
\alias{dmMapTerms}
\title{dmMapTerms: a data frame with 27 landmark symbols and definitions for cell fate map, 1986}
\format{data.frame}
\source{
\url{https://books.google.com/books?id=ol3\_CAAAQBAJ}
}
\usage{
dmMapTerms
}
\description{
dmMapTerms: a data frame with 27 landmark symbols and definitions for cell fate map, 1986
}
\keyword{datasets}
