% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exNmf21}
\alias{exNmf21}
\title{exNmf21: an NMFfit to the expressionPatterns of Dm blastoderm}
\format{\code{\link[NMF]{NMFfit-class}} instance}
\usage{
exNmf21
}
\description{
exNmf21: an NMFfit to the expressionPatterns of Dm blastoderm
}
\keyword{datasets}
