% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expressionPatterns}
\alias{expressionPatterns}
\title{expressionPatterns: 405 x 1640 data.frame of expression patterns for 701 genes with some replication}
\format{data.frame}
\source{
\url{http://www.pnas.org/content/113/16/4290.full} supplements
}
\usage{
expressionPatterns
}
\description{
expressionPatterns: 405 x 1640 data.frame of expression patterns for 701 genes with some replication
}
\keyword{datasets}
