% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_hierarchy.R
\name{fate_bias_summary}
\alias{fate_bias_summary}
\title{Plot summary of fate bias analysis}
\usage{
fate_bias_summary(fate_bias)
}
\arguments{
\item{fate_bias}{fata bias analysis result from function clone_fate_bias}
}
\description{
Plot summary of fate bias analysis
}
