% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_hierarchy.R
\name{fate_ct_col}
\alias{fate_ct_col}
\title{Column plot for Composition of cell types of each barcode}
\usage{
fate_ct_col(data, idx = "cell_type", barcode_use = NULL)
}
\arguments{
\item{data}{data.frame, indicating lineage tracing data, first column should
be lineage tracing barcodes, second column should be related cell type}

\item{idx}{column name of cell type}

\item{barcode_use}{Vector of barcodes to plot}
}
\value{
ggplot column plot
}
\description{
make column plot to show Composition of cell types of each barcode
}
\examples{
fate_ct_col(fate_test)
}
