% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_hierarchy.R
\name{fate_ct_heatmap}
\alias{fate_ct_heatmap}
\title{fate mapping heatmap}
\usage{
fate_ct_heatmap(
  data,
  idx = "cell_type",
  order_use = NULL,
  show_rownames = T,
  col = rev(colorRampPalette(c("#cc0000", "#FFff00", "#66ccff", "#000066"))(50)),
  cluster_row = F,
  cluster_cols = T,
  ...
)
}
\arguments{
\item{data}{data.frame, indicating lineage tracing data, first column should
be lineage tracing barcodes, second column should be related cell type}

\item{idx}{column name of cell type}

\item{order_use}{order cell type}

\item{show_rownames}{vector,indicating the order of cell types in the heatmap.}

\item{col}{heatmap colors}

\item{cluster_row}{boolean values determining if rows should be clustered}

\item{cluster_cols}{boolean values determining if cols should be clustered}

\item{...}{other pheatmap parameters}
}
\value{
heatmap
}
\description{
plot heatmap for lineage tracing barcodes in each cell type
}
\examples{
fate_ct_heatmap(fate_test)
}
