% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genebar.R
\name{genebar}
\alias{genebar}
\title{use a barplot to depict gene-oriented coefficients of a principal pattern}
\usage{
genebar(ppind = 1, depth = 30, coefdata)
}
\arguments{
\item{ppind}{principal pattern number: row index for 'sPPcoefGenes' to display}

\item{depth}{number of coefficients (starting at largest) to display}

\item{coefdata}{a data.frame with coefficients for genes}
}
\description{
use a barplot to depict gene-oriented coefficients of a principal pattern
}
