% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matops.R
\name{getXY}
\alias{getXY}
\title{convert matrix representing values on a grid to coordinates and values}
\usage{
getXY(mat, threshold)
}
\arguments{
\item{mat}{data matrix}

\item{threshold}{remove points with values less than threshold}
}
\description{
convert matrix representing values on a grid to coordinates and values
}
