% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppatt.R
\name{ggBlast}
\alias{ggBlast}
\title{use ggplot to render an image of a Drosophila blastocyst}
\usage{
ggBlast(basis, threshold = 3, template, peel = 0, ...)
}
\arguments{
\item{basis}{matrix of basis vectors for a reexpression of the spatial expression matrix}

\item{threshold}{numerical lower bound on basis elements to be included when convex hulls of points forming principal patterns are computed}

\item{template}{matrix of 0 and 1, 1 indicating which parts of image rectangle are represented in the basis}

\item{\dots}{not used}
}
\description{
use ggplot to render an image of a Drosophila blastocyst
}
\note{
Sharp thresholding can lead to warnings about patterns (vectors) that have no elelements larger than the threshold.
}
