% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matops.R
\name{matit}
\alias{matit}
\title{take an estimated basis vector from NMF and format to template tmpl}
\usage{
matit(x, nr = 16, nc = 32, tmpl)
}
\arguments{
\item{x}{numerical vector, in the drosmap application, has length 405 representing points in the blastocyst model ellipse}

\item{nr}{number of rows of output matrix}

\item{nc}{number of columns of output matrix}

\item{tmpl}{matrix of nr x nc that has 0 where cells do not exist; in drosmap application, sum(tmpl) == 405}
}
\description{
take an estimated basis vector from NMF and format to template tmpl
}
\examples{
require(NMF)
data(exNmf21)
data(template405)
bas = basis(exNmf21)
mm = matit(bas[,1], tmpl=template405)
contour(t(mm)) # first principal pattern using naive NMF
}
