% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planiTerms}
\alias{planiTerms}
\title{planiTerms: 17 landmark descriptions for the drosophila cell fate map}
\format{data.frame}
\source{
\url{http://www.sdbonline.org/sites/FLY/atlas/00atlas.htm}
}
\usage{
planiTerms
}
\description{
planiTerms: 17 landmark descriptions for the drosophila cell fate map
}
\keyword{datasets}
