% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomi.R
\name{plot_clone_embedding}
\alias{plot_clone_embedding}
\title{Title}
\usage{
plot_clone_embedding(
  barcode_use,
  pbmc,
  colors = c(rgb(200/255, 200/255, 200/255), rgb(230/255, 230/255, 230/255), rgb(239/255,
    153/255, 81/255), rgb(91/255, 166/255, 218/255), rgb(217/255, 83/255, 25/255))
)
}
\arguments{
\item{barcode_use}{vector, indicating barcodes for ploting}

\item{pbmc}{seurat object or dataframe}
}
\description{
Title
}
