% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PP}
\alias{PP}
\title{PP: Wu et al. PNAS 2016 principal patterns}
\format{data.frame (405 x 21)}
\source{
\url{http://www.pnas.org/content/113/16/4290.full} supplements
}
\usage{
PP
}
\description{
PP: Wu et al. PNAS 2016 principal patterns
}
\note{
see template405
}
\keyword{datasets}
