% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sPPcoefGenes}
\alias{sPPcoefGenes}
\title{sPPcoefGenes: 21 x 701 data.frame with selected gene-specific coefficients}
\format{data.frame}
\source{
\url{http://www.pnas.org/content/113/16/4290.full} supplements
}
\usage{
sPPcoefGenes
}
\description{
sPPcoefGenes: 21 x 701 data.frame with selected gene-specific coefficients
}
\keyword{datasets}
