% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sPPcoefPatterns}
\alias{sPPcoefPatterns}
\title{sPPcoefPatterns: 22 x 1640 data.frame with intercept and coefficients of NMF factorization of expressionPatterns}
\format{data.frame}
\source{
\url{http://www.pnas.org/content/113/16/4290.full} supplements
}
\usage{
sPPcoefPatterns
}
\description{
sPPcoefPatterns: 22 x 1640 data.frame with intercept and coefficients of NMF factorization of expressionPatterns
}
\keyword{datasets}
