% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template405}
\alias{template405}
\title{template405: 16 x 32 binary matrix, with zeroes outside the planimetric model (elliptical) for Drosophila blastoderm}
\format{matrix}
\usage{
template405
}
\description{
template405: 16 x 32 binary matrix, with zeroes outside the planimetric model (elliptical) for Drosophila blastoderm
}
\keyword{datasets}
