% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniqueGenes}
\alias{uniqueGenes}
\title{uniqueGenes: character vector of names of genes with spatially restricted expression in Drosophila}
\format{character vector}
\usage{
uniqueGenes
}
\description{
uniqueGenes: character vector of names of genes with spatially restricted expression in Drosophila
}
\keyword{datasets}
