% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict.R
\name{vizRestriction}
\alias{vizRestriction}
\title{render a view of spatial expression pattern of a gene}
\usage{
vizRestriction(edata, sym = "salm", tmpl = template405, threshold = 0.1)
}
\arguments{
\item{edata}{a matrix or data frame with columns representing genes and rows representing spatially organized measures of expression}

\item{sym}{gene symbol, a column name in edata}

\item{tmpl}{template as used in \code{\link{matit}}}

\item{threshold}{numeric value beneath which expression is regarded as absent}
}
\description{
render a view of spatial expression pattern of a gene
}
\examples{
data(expressionPatterns)
data(template405)
vizRestriction(expressionPatterns, sym="salm")
}
